New : user can write a note on each slide
We implanted the possibility of adding notes
to the presentation. In the editor, a textarea
can be filled with a text for each screen of the
current screencast. Those notes can be read in
the preview with a pluggin that can display/hide.

Added :
	- a 'note model'
	- a 'note layout'
	- a 'note view'

Note model only contains a single text and an Id (for
the DOM). Note View catches events related to changing
screen and saving, in order to load and save the notes.
Currently the note is saved on the textarea 'focusout'
and on the action save in the menu (wich doesn't
trigger 'focusout').

Modification :

	- bower.json : add dependencies to make the
pluggin note work with deckJS
	- Gruntfile.js : allow the copy of deck.js
files to the right location
	- Screen Model : include a 'note' attribute
	- Compiler et template Screencast : change
the access to the datas given by the model with
a .toJSON() function.
	- Dahuapp
	- Layout Workspace : add a note region to the
layout.